Moved library specific packages to pkg/ to account for new monorepo design
