+++ pkg/api/api.go
@@ -4,10 +4,10 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	credstackError "github.com/credstack/credstack-lib/errors"
-	"github.com/credstack/credstack-lib/header"
-	"github.com/credstack/credstack-lib/oauth/jwk"
-	"github.com/credstack/credstack-lib/server"
+	credstackError "github.com/credstack/credstack-lib/pkg/errors"
+	"github.com/credstack/credstack-lib/pkg/header"
+	"github.com/credstack/credstack-lib/pkg/oauth/jwk"
+	"github.com/credstack/credstack-lib/pkg/server"
 	"github.com/credstack/credstack-models/proto/api"
 	"go.mongodb.org/mongo-driver/v2/bson"
 	"go.mongodb.org/mongo-driver/v2/mongo"
