+++ pkg/application/application.go
@@ -4,10 +4,10 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	credstackError "github.com/credstack/credstack-lib/errors"
-	"github.com/credstack/credstack-lib/header"
-	"github.com/credstack/credstack-lib/secret"
-	"github.com/credstack/credstack-lib/server"
+	credstackError "github.com/credstack/credstack-lib/pkg/errors"
+	"github.com/credstack/credstack-lib/pkg/header"
+	"github.com/credstack/credstack-lib/pkg/secret"
+	"github.com/credstack/credstack-lib/pkg/server"
 	applicationModel "github.com/credstack/credstack-models/proto/application"
 	"go.mongodb.org/mongo-driver/v2/bson"
 	"go.mongodb.org/mongo-driver/v2/mongo"
