+++ pkg/oauth/flow/client_credentials.go
@@ -2,10 +2,10 @@ package flow
 
 import (
 	"crypto/subtle"
-	"github.com/credstack/credstack-lib/application"
-	"github.com/credstack/credstack-lib/oauth/claim"
-	"github.com/credstack/credstack-lib/oauth/token"
-	"github.com/credstack/credstack-lib/server"
+	"github.com/credstack/credstack-lib/pkg/application"
+	"github.com/credstack/credstack-lib/pkg/oauth/claim"
+	"github.com/credstack/credstack-lib/pkg/oauth/token"
+	"github.com/credstack/credstack-lib/pkg/server"
 	tokenModel "github.com/credstack/credstack-models/proto/token"
 )
