+++ pkg/oauth/jwk/revoke.go
@@ -4,8 +4,8 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	credstackError "github.com/credstack/credstack-lib/errors"
-	"github.com/credstack/credstack-lib/server"
+	credstackError "github.com/credstack/credstack-lib/pkg/errors"
+	"github.com/credstack/credstack-lib/pkg/server"
 	"go.mongodb.org/mongo-driver/v2/bson"
 	"go.mongodb.org/mongo-driver/v2/mongo"
 )
