+++ pkg/oauth/token/hs256.go
@@ -2,7 +2,7 @@ package token
 
 import (
 	"fmt"
-	"github.com/credstack/credstack-lib/secret"
+	"github.com/credstack/credstack-lib/pkg/secret"
 	tokenModel "github.com/credstack/credstack-models/proto/token"
 	"github.com/golang-jwt/jwt/v5"
 	pbTimestamp "google.golang.org/protobuf/types/known/timestamppb"
