+++ pkg/secret/rand.go
