+++ pkg/server/log.go
@@ -2,7 +2,7 @@ package server
 
 import (
 	"github.com/credstack/credstack-lib/internal"
-	"github.com/credstack/credstack-lib/options"
+	"github.com/credstack/credstack-lib/pkg/options"
 	"go.uber.org/zap"
 	"go.uber.org/zap/zapcore"
 	"os"
