+++ pkg/user/register.go
@@ -4,10 +4,10 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	credstackError "github.com/credstack/credstack-lib/errors"
-	"github.com/credstack/credstack-lib/header"
-	"github.com/credstack/credstack-lib/options"
-	"github.com/credstack/credstack-lib/server"
+	credstackError "github.com/credstack/credstack-lib/pkg/errors"
+	"github.com/credstack/credstack-lib/pkg/header"
+	"github.com/credstack/credstack-lib/pkg/options"
+	"github.com/credstack/credstack-lib/pkg/server"
 	userModel "github.com/credstack/credstack-models/proto/user"
 	"go.mongodb.org/mongo-driver/v2/bson"
 	"go.mongodb.org/mongo-driver/v2/mongo"
